app/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
decorator==5.1.1
demucs==4.0.1
fastapi==0.115.8
h11==0.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
pillow==10.4.0
proglog==0.1.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pytube==15.0.0
sniffio==1.3.1
starlette==0.45.3
supabase==2.13.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
youtube-dl==2021.12.17
yt-dlp==2025.1.26